requirements.txt
openai>=1.11.1
python-dotenv>=1.0.1
pillow>=10.0.0
requests>=2.31.0
